fix(nix): only start the database once when testing
To build and test the crate, we need a database to be running, which is
accomplished using `postgresqlStart` that we run in the `preBuild` hook.

The issue is that `postgresqlStart` is also added to `preCheck`, which
means we'll have to do some nasty hacks to avoid recreating the
database.

Another issue is that `cargoCheckHook` executes `preCheck`, which makes
overriding the hooks harder and leads to conflict.

A workaround is to remove `postgresqlStart` from `preCheckHooks`, which
gives us a better control on when to start the database.